requirements.txt
numpy
gensim
pymysql
Click==7.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
SQLAlchemy==1.3.11
Werkzeug==0.16.0
python-docx==0.8.10
tensorflow==1.14.0
keras==2.2.4
sklearn



